+++ database/seeds/InventoryItemsSeeder.php
@@ -14,10 +14,10 @@ class InventoryItemsSeeder extends Seeder
     {
         Schema::disableForeignKeyConstraints();
         DB::table('inventory_items')->truncate();
-        InventoryItem::create(['name' => 'Dixit', 'status_id' => 0]);
-        InventoryItem::create(['name' => 'Shadow Hunters', 'status_id' => 0]);
-        InventoryItem::create(['name' => 'Puissance 4 géant', 'status_id' => 1]);
-        InventoryItem::create(['name' => 'Wii U', 'status_id' => 2]);
+        InventoryItem::create(['name' => 'Dixit', 'status_id' => 1]);
+        InventoryItem::create(['name' => 'Shadow Hunters', 'status_id' => 1]);
+        InventoryItem::create(['name' => 'Puissance 4 géant', 'status_id' => 2]);
+        InventoryItem::create(['name' => 'Wii U', 'status_id' => 3]);
         Schema::enableForeignKeyConstraints();
     }
 }